Loader/requirements.txt
Pillow>=10.0.0
pycryptodome>=3.20.0